Add files via upload
Login Page: was modified so that the login credentials are checked against the local database, therefore only reroutes to the Garage if the username and password match "our" records
MainViewModel is used and I deleted some of the content that we weren't currently using just to make sure that once we need that other code we can annotate it properly. 
My Garage: was modified so that the list of cars reflects what is shown in the local database. (One flaw is that it currently shows entire Garage table, so all users' vehicles. Currently having a hard time passing the username from page to page since databinding doesn't work that way I found out during this mod, we will need to adjust that once we figure out how to (thumbs up))
Registration Page: was modified so that it no longer breaks when an new user attempts to input a username that has already been used. Instead it gives a warning and then allows redirects them back to a new registration page.
Vehicle Registration Page: was modified so that it no longer breaks when a user tries to input a vehicle name that has already been previously used (Again this isn't to check with the association with the username so we will have to modify this query once we figure out how to pass in the username (thumbs up)).